Restructure to use `/api` folder with `guestId` (#6)
* Change to api folder structure

* Change from id to guestId

* Refactor folder structure and imports
