+++ app/api/guests+api.ts
@@ -1,6 +1,6 @@
-import { createGuestInsecure, getGuestsInsecure } from '../database/guests';
-import { ExpoApiResponse } from '../ExpoApiResponse';
-import { Guest, guestsSchema } from '../migrations/00000-createTableGuests';
+import { createGuestInsecure, getGuestsInsecure } from '../../database/guests';
+import { ExpoApiResponse } from '../../ExpoApiResponse';
+import { Guest, guestsSchema } from '../../migrations/00000-createTableGuests';
 
 type ExpoResponseBodyGet = {
   guests: Guest[];
