+++ components/UserItem.tsx
@@ -1,6 +1,5 @@
 import { StyleSheet, Text, View } from 'react-native';
 import { colors } from '../constants/colors';
-import { Guest } from '../migrations/00000-createTableGuests';
 
 const styles = StyleSheet.create({
   center: {
